Множество изменений. Минимально рабочий продукт

+++ data/src/main/java/com/example/data/local/entity/CalendarEntity.kt
@@ -6,16 +6,10 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "calendar_habits")
 data class CalendarEntity(
-    @PrimaryKey(autoGenerate = true)
-    var calendarId: Long = 0L,
-    @ColumnInfo(name = "calendar_habit_id")
-    var calendarHabitId: Long = 0L,
-    @ColumnInfo(name = "calendar_habit_price")
-    var calendarHabitPrice: Int = 0,
-    @ColumnInfo(name = "calendar_habit_quantity")
-    var calendarHabitQuantity: Float = 0F,
-    @ColumnInfo(name = "calendar_habit_description")
-    var calendarHabitDescription: String = "",
-    @ColumnInfo(name = "calendar_date")
-    var calendarDate: String = "2010-01-01"
+    @PrimaryKey(autoGenerate = true) var calendarId: Long = 0L,
+    @ColumnInfo(name = "calendar_habit_id") var calendarHabitId: Long = 0L,
+    @ColumnInfo(name = "calendar_habit_price") var calendarHabitPrice: Float = 0F,
+    @ColumnInfo(name = "calendar_habit_quantity") var calendarHabitQuantity: Float = 0F,
+    @ColumnInfo(name = "calendar_habit_description") var calendarHabitDescription: String = "",
+    @ColumnInfo(name = "calendar_date") var calendarDate: String = "2010-01-01"
 )
